feat(minifyJs): fork minifyJs child processes
