+++ scripts/build-sys-node.js
@@ -12,6 +12,10 @@ const DEST_FILE = path.join(__dirname, '..', 'dist', 'sys', 'node', 'index.js');
 
 const success = transpile(path.join('..', 'src', 'sys', 'node', 'tsconfig.json'));
 
+const whitelist = [
+  'uglify-es'
+];
+
 if (success) {
   bundle('node-fetch.js');
   bundle('sys-util.js');
@@ -27,6 +31,21 @@ if (success) {
         libraryTarget: 'commonjs'
       },
       target: 'node',
+      externals: function(context, request, callback) {
+        if (request.match(/^(\.{0,2})\//)) {
+          // absolute and relative paths are not externals
+          return callback();
+        }
+
+        if (whitelist.indexOf(request) > -1) {
+          // we specifically do not want to bundle these imports
+          require.resolve(request);
+          return callback(null, request);
+        }
+
+        // bundle this import
+        callback();
+      },
       optimization: {
         minimize: false
       }
